# order-service/src/main/resources/openapi/order-service-api.yaml
openapi: 3.1.0
info:
  title: Order Service API
  description: Order lifecycle management — PostgreSQL · Kafka producer · port 8082
  version: "1.0.0"

servers:
  - url: /api/v1

tags:
  - name: Orders
    description: Order lifecycle operations

paths:

  /orders:
    get:
      tags: [Orders]
      operationId: getAllOrders
      summary: List all orders (admin only — requires products:write scope)
      responses:
        "200":
          description: All orders across all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
        "403":
          description: Access denied — caller does not have admin scope
    post:
      tags: [Orders]
      operationId: createOrder
      summary: Place a new order (PENDING preview)
      description: |
        Creates a new PENDING order for the currently authenticated user.
        Does NOT reserve stock and does NOT publish a Kafka event.
        The cart is NOT cleared. Call `POST /orders/{id}/confirm` to confirm and reserve stock.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        "201":
          description: Order placed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        "400":
          description: Validation error
        "401":
          description: Unauthorized

  /orders/{id}/confirm:
    post:
      tags: [Orders]
      operationId: confirmOrder
      summary: Confirm a PENDING order (reserve stock → CONFIRMED → publish Kafka event)
      description: |
        Verifies that the order is PENDING and owned by the caller, then calls
        product-service to atomically reserve stock. On success the order transitions
        PENDING → CONFIRMED and an `order.confirmed.v1` Kafka event is published.
        Returns 409 if stock is insufficient (order stays PENDING, cart stays intact).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        "403":
          description: Caller is not the order owner
        "404":
          description: Order not found
        "409":
          description: Insufficient stock for one or more items

  /orders/{id}:
    get:
      tags: [Orders]
      operationId: getOrderById
      summary: Get order by ID (owner or service account)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        "403":
          description: Access denied — caller is not the order owner
        "404":
          description: Order not found

  /orders/{id}/status:
    put:
      tags: [Orders]
      operationId: updateOrderStatus
      summary: Update order status (admin / internal)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequest'
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        "400":
          description: Validation error
        "404":
          description: Order not found

  /orders/user/{userId}:
    get:
      tags: [Orders]
      operationId: getOrdersByUserId
      summary: Get all orders for a user (owner or service account)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
        "403":
          description: Access denied — caller is not the requested user

components:
  schemas:

    CreateOrderRequest:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: |
            Internal user UUID (from user-service). Required when the request is made by a
            service account (M2M Client Credentials token). Ignored when called by an end-user
            JWT — the userId is resolved from the token's `sub` claim in that case.
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemRequest'

    OrderItemRequest:
      type: object
      properties:
        productId:
          type: string
          description: Product ID (MongoDB ObjectId as hex string)
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          format: double
          minimum: 0.0

    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/OrderStatus'

    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        totalAmount:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    OrderItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          description: Product ID (MongoDB ObjectId as hex string)
        quantity:
          type: integer
        unitPrice:
          type: number
          format: double

    OrderStatus:
      type: string
      enum: [PENDING, CONFIRMED, SHIPPED, DELIVERED, CANCELLED]